trap: implement following fixes for aarch64:
1. Accesses were being made to *_EL3 registers in
   exception code. qemu doesn't implement these registers
   and our kernel only cares about running at EL1. Fix this.
2. The structure of struct trapframe was out of date from
   its usage in trap_asm.S. Fix this. Note pc --> elr
3. Add missing add instruction to shift sp in trap_asm.S.

+++ xv6-armv8/arm.c
@@ -20,7 +20,7 @@ void sti (void)
 // return whether interrupt is currently enabled
 int int_enabled ()
 {
-    uint32 val;
+    uint64 val;
 
     asm("MRS %[v], DAIF": [v]"=r" (val)::);
